Abstracted the batch processing algorithm
